+++ packages/api-queries/src/index.ts
@@ -14,7 +14,6 @@ export * from './domain';
 export * from './domains';
 export * from './emails';
 export * from './geo';
-export * from './marketplace-products';
 export * from './me-a8c';
 export * from './me-account';
 export * from './me-account-recovery';
@@ -29,6 +28,7 @@ export * from './me-two-step';
 export * from './meta-sms-country-codes';
 export * from './p2';
 export * from './performance';
+export * from './plugin';
 export * from './purchase';
 export * from './site-activity-log';
 export * from './site-address-change';
@@ -75,4 +75,3 @@ export * from './site';
 export * from './sites';
 export * from './ssh';
 export * from './plugins';
-export * from './wp-org-plugin';
