--- packages/api-queries/src/marketplace-products.ts
@@ -1,8 +1,0 @@
-import { fetchMarketplacePlugin } from '@automattic/api-core';
-import { queryOptions } from '@tanstack/react-query';
-
-export const wooCommercePluginQuery = ( slug: string ) =>
-	queryOptions( {
-		queryKey: [ 'marketplace-products', slug ],
-		queryFn: () => fetchMarketplacePlugin( slug ),
-	} );
